change arithmetic operations to accept double values

+++ src/commands/Adder.java
@@ -4,9 +4,9 @@ public class Adder implements Calculation {
 
 	@Override
 	public String process(String[] args) {
-		int result = Integer.parseInt(args[1]);
+		double result = Integer.parseInt(args[1]);
 		for (int i = 2; i < args.length; i++) {
-			result += Integer.parseInt(args[i]);
+			result += Double.parseDouble(args[i]);
 		}
 		return String.valueOf(result);
 	}
